refactor: rename prediction_leaderbiard dir into src
