+++ .gitignore
@@ -205,4 +205,4 @@ cython_debug/
 marimo/_static/
 marimo/_lsp/
 __marimo__/
-prediction_leaderboard/db/
+src/db/
