+++ src/config.yaml
@@ -1,0 +1,5 @@
+public_ground_truth: ./data/public_gt.csv
+private_ground_truth: ./data/private_gt.csv
+default_leaderboard_limit: 10
+default_history_limit: 10
+metric: rmse
