+++ src/submit.py
@@ -1,12 +1,15 @@
 import argparse
 import os
 import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+from pathlib import Path
 import sqlite3
 import yaml
 import pandas as pd
 from datetime import datetime
-from prediction_leaderboard.evaluator import metrics as metric_lib
+
+os.chdir(Path(__file__).parent)
+
+from evaluator import metrics as metric_lib
 
 CONFIG_PATH = os.path.join(os.path.dirname(__file__), 'config.yaml')
 DB_PATH = os.path.join(os.path.dirname(__file__), 'db', 'leaderboard.sqlite')
